% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_series.R
\name{add_suitability_stars}
\alias{add_suitability_stars}
\title{Estimate suitability across space and add it to a stars object}
\usage{
add_suitability_stars(stars_object, min_years_trigger_suitability = 10)
}
\arguments{
\item{stars_object}{a stars object}

\item{min_years_trigger_suitability}{the minimum number of suitable winters to trigger the establishment or the disappearance of bats in a region}
}
\value{
a stars object
}
\description{
Estimate suitability across space and add it to a stars object
}
\examples{
data_OBSCLIM <- readRDS("../NC/stars_winter/gswp3-w5e5_OBSCLIM_winter.rds")
add_suitability_stars(data_OBSCLIM)

}
