% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{arguments}
\alias{arguments}
\title{Definition of the function arguments}
\arguments{
\item{base_size}{base font size, given in pts}

\item{clean}{a boolean indicating if the data need to be filter by \code{\link[=clean_Tskin_table]{clean_Tskin_table()}} or not (default = \code{TRUE})}

\item{data_budget}{a dataframe produced by \code{\link[=compute_nrg_budget]{compute_nrg_budget()}}}

\item{data_MR}{the data used to fit the thermoregulatory curves as produced by \code{\link[=build_MR_table]{build_MR_table()}}}

\item{data_temp}{a dataframe with at least columns \code{Date} and \code{Temp}, such as those produced by \code{\link[=build_temp_2years]{build_temp_2years()}}}

\item{data_Tskin}{a file produced by \code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}}

\item{filename}{the path to the file from which the data are to be imported}

\item{filenames}{a vector of paths to the files from which the data are to be imported}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}

\item{rangeTa}{the range of ambient temperature to consider in plot}

\item{rangeTskin}{the range of skin temperature to consider in plot}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 24)}

\item{Tmirror}{the ambient temperature around which predictions are mirrored}

\item{y}{a string of characters indicating what y-variable to plot:
"g_fat_per_state", "g_fat_per_day", or "g_fat_per_winter"}
}
\description{
Here is the list of most function arguments used in the package.
The arguments not contained in this list are those for which the exact meaning depends on the context.
}
