% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kharkiv.R
\name{build_Kharkiv_table}
\alias{build_Kharkiv_table}
\title{Build table with raw temperature data from Kharkiv, Ukraine}
\usage{
build_Kharkiv_table(filename)
}
\arguments{
\item{filename}{the name of the data to be imported}
}
\value{
a dataframe
}
\description{
This function loads the file containing temperature data from years 2011 and 2012 in Kharkiv, Ukraine.
}
\examples{
filepath <- list.files(system.file("extdata/weather_real", package = "winteR"),
                       full.names = TRUE)[1]
data_Kharkiv <- build_Kharkiv_table(filepath)
head(data_Kharkiv)

}
