% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoreg.R
\name{build_MR_table}
\alias{build_MR_table}
\title{Build table with raw physiological data on metabolic rate}
\usage{
build_MR_table(filename)
}
\arguments{
\item{filename}{the path to the file from which the data are to be imported}
}
\value{
a dataframe
}
\description{
This function loads the file containing physiological data that will be used to fit the
thermoregulatory curves and it converts the metabolic rate in various units. It should be
structured as follows: TODO Kseniia, please add description XXX .
}
\examples{
filepath <- paste0(system.file("extdata/thermoreg", package = "winteR"),
                   "/metabolic_rate.txt")
data_MR <- build_MR_table(filepath)
head(data_MR)

}
