% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{build_physio_table}
\alias{build_physio_table}
\title{Build table with raw physiological data}
\usage{
build_physio_table(files, clean = TRUE)
}
\arguments{
\item{files}{a vector of paths to the files storing the physiological data}

\item{clean}{a boolean indicating if the data need to be filter by \code{\link[=clean_physio_table]{clean_physio_table()}} or not (default = \code{TRUE})}
}
\value{
a dataframe
}
\description{
This function creates the table of raw physiological data.
}
\examples{
files_to_do <- list.files(system.file("extdata/physio", package = "winteR"), full.names = TRUE)
data_physio <- build_physio_table(files_to_do)
head(data_physio)

}
\seealso{
\code{\link[=load_physio_datafile]{load_physio_datafile()}}, \code{\link[=classify_physio_state]{classify_physio_state()}}, \code{\link[=clean_physio_table]{clean_physio_table()}}
}
