% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNCfiles.R
\name{build_source_stars}
\alias{build_source_stars}
\title{Create stars object from NC files and store them on hard drive as RDS files}
\usage{
build_source_stars(
  metadirectory_NCfiles,
  directory_stars,
  downsample = FALSE,
  crop = TRUE,
  .downsampling_args = c(5, 5, 0),
  .crop_args = list(S = 27, N = 72, W = -13, E = 56),
  nb_cores = 2
)
}
\arguments{
\item{metadirectory_NCfiles}{a folder containing one subfolder per climate model, themselves containing one subfolder per SSP scenarios, themsevles containing the NC files}

\item{directory_stars}{the destination location for the rds files storing the stars object}

\item{downsample}{a boolean indicating whether to reduce the resolution or not (default = \code{FALSE})}

\item{crop}{a boolean indicating whether to crop or not (default = \code{TRUE})}

\item{.downsampling_args}{arguments for \code{\link[stars:st_downsample]{stars::st_downsample()}} (default = \code{c(5, 5, 0)}, i.e. aggregate space but not dates)}

\item{.crop_args}{the list of coordinates to use for cropping (default = \code{list(S = 27, N = 72, W = -13, E = 56)})}

\item{nb_cores}{the number of CPU cores to use (for Linux or MacOS only, not Windows, and don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}
}
\value{
nothing (creates files on hard drive)
}
\description{
This function is used to create all stars objects storing the daily temperature data, which is
required to compute the hibernation niche. We do not provide the raw NC files since they are
enormous, but we downloaded them from \href{https://www.isimip.org/}{the Inter-Sectoral Impact Model Intercomparison Project} using the python script we wrote (look for
download_NC_files_isimip.py). The interface to ISIMIP may change, so please refer to instruction
on \url{https://www.isimip.org/}.
}
\examples{
#See ?winteR

}
\seealso{
\code{\link[=load_NC_file]{load_NC_file()}}, \code{\link[=load_NC_files]{load_NC_files()}}
}
