% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{build_stars_winter.stats}
\alias{build_stars_winter.stats}
\title{Build a stars object with summary statistics aggregated across years}
\usage{
build_stars_winter.stats(stars_object)
}
\arguments{
\item{stars_object}{a stars object}
}
\value{
a stars object
}
\description{
Build a stars object with summary statistics aggregated across years
}
\examples{
data_OBSCLIM <- readRDS("../NC/stars_winter/gswp3-w5e5_OBSCLIM_winter.rds")
build_stars_winter.stats(data_OBSCLIM)

}
