% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_series.R
\name{build_suitability_stars}
\alias{build_suitability_stars}
\title{Estimate suitability across space}
\usage{
build_suitability_stars(directory_stars, min_years_trigger_suitability = 5)
}
\arguments{
\item{directory_stars}{the destination location for the rds files storing the stars object}

\item{min_years_trigger_suitability}{the minimum number of suitable winters to trigger the establishment or the disappearance of bats in a region}
}
\value{
a stars object
}
\description{
Estimate suitability across space
}
\examples{
run <- FALSE
if (run) {
  build_suitability_stars("../NC/stars_winter")
}

}
