% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_series.R
\name{build_temp_2years}
\alias{build_temp_2years}
\title{Build table with raw temperature data}
\usage{
build_temp_2years(filename)
}
\arguments{
\item{filename}{the path to the file from which the data are to be imported}
}
\value{
a dataframe
}
\description{
This function loads a file containing date and temperature data for 2 consecutive years-
}
\examples{
file_Kharkiv <- paste0(system.file("extdata/weather_real", package = "winteR"),
                       "/Kharkiv_weather_2011_2012.csv")
data_Kharkiv <- build_temp_2years(file_Kharkiv)
head(data_Kharkiv)

}
