% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tskin_normothermy.R
\name{build_Tskin_table}
\alias{build_Tskin_table}
\title{Build table with raw physiological data on skin temperature}
\usage{
build_Tskin_table(files, clean = TRUE)
}
\arguments{
\item{files}{a vector of paths to the files storing the physiological data}

\item{clean}{a boolean indicating if the data need to be filter by \code{\link[=clean_Tskin_table]{clean_Tskin_table()}} or not (default = \code{TRUE})}
}
\value{
a dataframe
}
\description{
This function creates the table of raw physiological data.
}
\examples{
files_to_do <- list.files(system.file("extdata/Tskin", package = "winteR"), full.names = TRUE)
data_Tskin <- build_Tskin_table(files_to_do)
head(data_Tskin)

}
\seealso{
\code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}, \code{\link[=classify_Tskin_state]{classify_Tskin_state()}}, \code{\link[=clean_Tskin_table]{clean_Tskin_table()}}
}
