% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checkDependencies}
\alias{checkDependencies}
\title{Check optional dependencies}
\usage{
checkDependencies()
}
\value{
a boolean indicating if all dependencies are met or not
}
\description{
This function checks that the dependencies not required to install the package but needed to run
the workflow are all met.
}
\examples{
checkDependencies()

}
