% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{classify_physio_state}
\alias{classify_physio_state}
\title{Classify the physiological state of individuals}
\usage{
classify_physio_state(data)
}
\arguments{
\item{data}{a file created with \code{\link[=load_physio_datafile]{load_physio_datafile()}}}
}
\value{
a vector with the physiological state at each time point
}
\description{
This function classifies the physiological state based on the difference between the body temperature
and the ambient temperature. If $Tb > Ta + ((max(Tb) - Ta) / 2)$ the bat is considered in normothermy,
otherwise, it is considered to be in torpor.
}
\details{
This function should not be directly called by the user.
It is called internally when calling \code{\link[=build_physio_table]{build_physio_table()}}.
}
\examples{
filespath <- list.files(system.file("extdata/physio", package = "winteR"), full.names = TRUE)
data_physio <- load_physio_datafile(filespath[1])
data_physio$State <- classify_physio_state(data_physio)
head(data_physio)

}
\seealso{
\code{\link[=load_physio_datafile]{load_physio_datafile()}}, \code{\link[=build_physio_table]{build_physio_table()}}, \code{\link[=clean_physio_table]{clean_physio_table()}}
}
