% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tskin_normothermy.R
\name{classify_Tskin_state}
\alias{classify_Tskin_state}
\title{Classify the physiological state of individuals based on skin temperature}
\usage{
classify_Tskin_state(data)
}
\arguments{
\item{data}{a file created with \code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}}
}
\value{
a vector with the physiological state at each time point
}
\description{
This function classifies the physiological state based on the difference between the skin temperature
and the ambient temperature. If $Tskin > Ta + ((max(Tskin) - Ta) / 2)$ the bat is considered in normothermy,
otherwise, it is considered to be in torpor.
}
\details{
This function should not be directly called by the user.
It is called internally when calling \code{\link[=build_Tskin_table]{build_Tskin_table()}}.
}
\examples{
filespath <- list.files(system.file("extdata/Tskin", package = "winteR"), full.names = TRUE)
data_physio <- load_Tskin_datafile(filespath[1])
data_physio$State <- classify_Tskin_state(data_physio)
head(data_physio)

}
\seealso{
\code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}, \code{\link[=build_Tskin_table]{build_Tskin_table()}}, \code{\link[=clean_Tskin_table]{clean_Tskin_table()}}
}
