% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{clean_physio_table}
\alias{clean_physio_table}
\title{Clean up the physiological data}
\usage{
clean_physio_table(data)
}
\arguments{
\item{data}{the dataframe produced by \code{\link[=build_physio_table]{build_physio_table()}}}
}
\value{
a dataframe
}
\description{
This function removes problematic observations of the physiological state; that is, observations
that seem to have happened after the death of the individual and all observations corresponding
to the drop of a sensor.
}
\details{
This function should not be directly called by the user.
It is called internally when calling \code{\link[=build_physio_table]{build_physio_table()}}.
}
\examples{
files_to_do <- list.files(system.file("extdata/physio", package = "winteR"), full.names = TRUE)
data_physio <- build_physio_table(files_to_do, clean = FALSE)
dim(data_physio)
data_physio_cleaned <- clean_physio_table(data_physio)
dim(data_physio_cleaned)

}
\seealso{
\code{\link[=load_physio_datafile]{load_physio_datafile()}}, \code{\link[=classify_physio_state]{classify_physio_state()}}, \code{\link[=build_physio_table]{build_physio_table()}}
}
