% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tskin_normothermy.R
\name{clean_Tskin_table}
\alias{clean_Tskin_table}
\title{Clean up the physiological data on skin temperature}
\usage{
clean_Tskin_table(data)
}
\arguments{
\item{data}{the dataframe produced by \code{\link[=build_Tskin_table]{build_Tskin_table()}}}
}
\value{
a dataframe
}
\description{
This function removes problematic observations of the physiological state; that is, observations
that seem to have happened after the death of the individual and all observations corresponding
to the drop of a sensor.
}
\details{
This function should not be directly called by the user.
It is called internally when calling \code{\link[=build_Tskin_table]{build_Tskin_table()}}.
}
\examples{
files_to_do <- list.files(system.file("extdata/Tskin", package = "winteR"), full.names = TRUE)
data_Tskin <- build_Tskin_table(files_to_do, clean = FALSE)
dim(data_Tskin)
data_Tskin_cleaned <- clean_Tskin_table(data_Tskin)
dim(data_Tskin_cleaned)

}
\seealso{
\code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}, \code{\link[=classify_Tskin_state]{classify_Tskin_state()}}, \code{\link[=build_Tskin_table]{build_Tskin_table()}}
}
