% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{compute_budget_stars}
\alias{compute_budget_stars}
\title{Compute budget for all winters from a stars object}
\usage{
compute_budget_stars(
  stars_object,
  fit_state,
  fit_MR,
  roost_insulation_dTa = 5,
  temp_threshold = 7,
  split_summer = "07-01",
  min_days_trigger_winter = 14,
  threshold_mortality = 27,
  nb_cores_location = 1,
  nb_cores_year = 1
)
}
\arguments{
\item{stars_object}{a stars object containing daily temperatures}

\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 24)}

\item{nb_cores_location}{the number of CPU cores to use to parallelize the same task across multiple locations (don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}

\item{nb_cores_year}{the number of CPU cores to use to parallelize the same task across multiple years (don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}
}
\value{
a stars object with energy budget information
}
\description{
Compute budget for all winters from a stars object
}
\examples{
run <- FALSE
if (run) {
  path_NC_dir <- "../NC/ISIMIP_sources/gswp3-w5e5/OBSCLIM/"
  obsclim <- load_NC_files(path_NC_dir, downsample = FALSE, .downsampling_args = c(40, 40, 0))
  obsclim1901_1904 <- dplyr::filter(obsclim, time < as.Date("1905-01-01"))
  test <- compute_budget_stars(obsclim1901_1904, fit_state = fit_normo_cauchit, fit_MR = fit_torpor, nb_cores_year = 5)
 }

}
