% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{compute_flightpower_excess}
\alias{compute_flightpower_excess}
\title{Estimate the excess power that an average common noctule of a given mass has for flying}
\usage{
compute_flightpower_excess(mass_g = 26.5, output4optim = FALSE)
}
\arguments{
\item{mass_g}{the mass of a common noctule}

\item{output4optim}{a boolean indicating if the function output must be modifed to direct use of optimization routines}
}
\value{
a data frame with various information about the power and ability of the bat to fly, unless \code{output4optim = TRUE}
}
\description{
This function is used to determines if an average common noctule can fly or not.
If the excess power is positive, it can fly, otherwise, it cannot.
If the \code{output4optim} parameter is set to \code{TRUE}, the function report the absolute excess power, which allows for searching the maximum mass that a noctule can reach and still be able to fly (see example).
}
\examples{
compute_flightpower_excess() # avg noctule can fly
compute_flightpower_excess(50) # 50 g noctule cannot
\dontrun{
optimize(compute_flightpower_excess, lower = 0, upper = 100,
         output4optim = TRUE)$minimum # max mass able to fly
}

}
