% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudes.R
\name{compute_northward_move}
\alias{compute_northward_move}
\title{Compute the distance in km between two latitudes}
\usage{
compute_northward_move(lat, lat_ref = 0, signed = TRUE)
}
\arguments{
\item{lat}{a latitude}

\item{lat_ref}{a latitude of reference}

\item{signed}{whether the output value should be signed or absolute (default = \code{TRUE}, which implies it is signed)}
}
\value{
a distance in km
}
\description{
Compute the distance in km between two latitudes
}
\examples{
## Earth circumference through the poles should be 40,007.863 km
compute_northward_move(90, -90)*2 # very close!

}
