% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{compute_nrg_budget}
\alias{compute_nrg_budget}
\title{Compute the energy budget associated to given ambient temperatures}
\usage{
compute_nrg_budget(data, fit_state, fit_MR, roost_insulation_dTa = 5)
}
\arguments{
\item{data}{a dataframe with at least columns \code{Date} and \code{Temp}}

\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}
}
\value{
a dataframe
}
\description{
Bugets are computed in grams of fat consumed.
}
\examples{
#See ?winteR

}
