% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_series.R
\name{compute_suitability_summarystats}
\alias{compute_suitability_summarystats}
\title{Extract summary statistics about winter suitability}
\usage{
compute_suitability_summarystats(
  vec_Suitable_winter,
  vec_Year,
  min_years_trigger_suitability = 5
)
}
\arguments{
\item{vec_Suitable_winter}{a vector of boolean indicating if a hibernation season was suitable (\code{TRUE}) or not (\code{FALSE}) (nb: \code{FALSE} includes lack of hibernation season as well as hibernation season where the budget is not sufficient to survive)}

\item{vec_Year}{a vector of years}

\item{min_years_trigger_suitability}{the minimum number of suitable winters to trigger the establishment or the disappearance of bats in a region}
}
\value{
a data frame
}
\description{
This function returns how often an average bat experience a suitable winter in a given location.
It also computes the year of the establishment of bats, as well as the year of their
disappearance in a given location. We consider that the year of establishment corresponds to the
first year within a sequence of at least 5 years in a row  (or another number as defined via the
parameter \code{min_years_trigger_suitability}) for which an hibernation season exists and an average
bat is expected to survive each such season. We consider that the year of disappearance
corresponds to the year 5 (i.e. \code{min_years_trigger_suitability}) within a sequence of at least 5
years in a row (i.e. \code{min_years_trigger_suitability}) for which either an hibernation season does
not exist, or an average bat is not expected to survive each such years.
}
\examples{
suitable_winters_vec <- c(rep(FALSE, 15), rep(TRUE, 4), FALSE, rep(TRUE, 6),
                          rep(FALSE, 10), rep(TRUE, 5), rep(TRUE, 2))
#suitable_winters_vec <- rep(TRUE, 10)
#suitable_winters_vec <- rep(FALSE, 10)
years <- seq(2000, length.out = length(suitable_winters_vec))
data.frame(Year = years, Suitable_winter = suitable_winters_vec)
compute_suitability_summarystats(vec_Suitable_winter = suitable_winters_vec, vec_Year = years)

}
