% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{compute_winter_summarystats}
\alias{compute_winter_summarystats}
\title{Extract summary statistics about winter characteristics}
\usage{
compute_winter_summarystats(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a data frame
}
\description{
Extract summary statistics about winter characteristics
}
\examples{
data_OBSCLIM <- readRDS("../NC/stars_winter/gswp3-w5e5_OBSCLIM_winter.rds")
stars::st_apply(data_OBSCLIM[c("Duration_winter", "Temp_winter_mean")],
                MARGIN = c("x", "y"),
                FUN = compute_winter_summarystats)

}
