% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{extract_stars_latitude}
\alias{extract_stars_latitude}
\title{Extract latitude from a stars object}
\usage{
extract_stars_latitude(
  stars_object,
  year = NULL,
  name_bool_var = "Survive",
  mask = NULL,
  fn = "median"
)
}
\arguments{
\item{stars_object}{a stars object}

\item{year}{a numeric scalar indicating the year of the winter to consider}

\item{name_bool_var}{the quoted name of a boolean variable to use to filter cells in the grid (default is "Survive")}

\item{mask}{a MULTIPOLYGON object to be used as a mask}

\item{fn}{a function to be applied}
}
\value{
the latitude
}
\description{
This function should not be called by the user. It is called internally by
\code{\link[=summarise_info_winter.stars]{summarise_info_winter.stars()}}.
}
\examples{
run <- FALSE
if (run) {
 test_stars <- readRDS("../NC/stars_winter/gfdl-esm4_SSP126_winter.rds")
 test_stars_small <- dplyr::filter(test_stars, year > 2080, x < -10, y > 70)
 extract_stars_latitude(test_stars_small)
 extract_stars_latitude(test_stars_small, fn = "max")
 }

}
