% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoreg.R
\name{filter_MR_table}
\alias{filter_MR_table}
\title{Keep physiological data corresponding to lowest metabolic rate per ambient temperature}
\usage{
filter_MR_table(data_MR, fit_MR)
}
\arguments{
\item{data_MR}{the data used to fit the thermoregulatory curves as produced by \code{\link[=build_MR_table]{build_MR_table()}}}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}
}
\value{
a dataframe
}
\description{
Keep physiological data corresponding to lowest metabolic rate per ambient temperature
}
\examples{
filepath <- paste0(system.file("extdata/thermoreg", package = "winteR"),
                   "/metabolic_rate.txt")
data_MR <- build_MR_table(filepath)
nrow(data_MR)
data_MR_filtered <- filter_MR_table(data_MR, fit_torpor)
nrow(data_MR_filtered)

}
