% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_series.R
\name{find_winter_temp2years}
\alias{find_winter_temp2years}
\title{Classify dates within 2 consecutive years as belonging to biological winter or not}
\usage{
find_winter_temp2years(
  data,
  split_summer = "07-01",
  temp_threshold = 7,
  min_days_trigger_winter = 14
)
}
\arguments{
\item{data}{a dataframe with at least columns \code{Date} and \code{Temp}}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}
}
\value{
a vector of boolean
}
\description{
Classify dates within 2 consecutive years as belonging to biological winter or not
}
\examples{
file_Kharkiv <- paste0(system.file("extdata/weather_real", package = "winteR"),
                       "/Kharkiv_weather_2011_2012.csv")
data_Kharkiv <- build_Kharkiv_table(file_Kharkiv)
find_winter_temp2years(data_Kharkiv)

}
