% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_torpor}
\alias{fit_torpor}
\title{Fitted thermoregulatory curves}
\format{
\subsection{\code{fit_torpor}}{

A list with two elements
\describe{
\item{params}{parameters estimated by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}
\item{ppo}{posteriors overlap values}
}
}
}
\usage{
fit_torpor
}
\description{
A model fitted with \code{\link[torpor:tor_fit]{torpor::tor_fit()}}.
}
\keyword{datasets}
