% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_overs.R
\name{is_winter}
\alias{is_winter}
\title{Is this date winter?}
\usage{
is_winter(x, start = 10, end = 3)
}
\arguments{
\item{x}{a date-time object}

\item{start}{month of winter start}

\item{end}{month of winter end}
}
\value{
logical
}
\description{
Is this date winter?
}
