% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoreg.R
\name{load_MR_datafile}
\alias{load_MR_datafile}
\title{Load the file containing physiological data on metabolic rate}
\usage{
load_MR_datafile(filename)
}
\arguments{
\item{filename}{the name of the data to be imported}
}
\value{
a dataframe
}
\description{
This function loads the file containing physiological data that will be used to fit the thermoregulatory curves.
It should be structured as follows: TODO Kseniia, please add description XXX .
}
\examples{
filepath <- list.files(system.file("extdata/thermoreg", package = "winteR"), full.names = TRUE)[1]
data_MR <- load_MR_datafile(filepath)
head(data_MR)

}
