% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNCfiles.R
\name{load_NC_file}
\alias{load_NC_file}
\title{Loads one temperature prediction grid stored as NC file}
\usage{
load_NC_file(
  filename,
  downsample = FALSE,
  crop = TRUE,
  .downsampling_args = c(5, 5, 1),
  .crop_args = list(S = 27, N = 72, W = -13, E = 56)
)
}
\arguments{
\item{filename}{the path to the file from which the data are to be imported}

\item{downsample}{a boolean indicating whether to reduce the resolution or not (default = \code{FALSE})}

\item{crop}{a boolean indicating whether to crop or not (default = \code{TRUE})}

\item{.downsampling_args}{arguments for \code{\link[stars:st_downsample]{stars::st_downsample()}} (default = \code{c(5, 5, 1)})}

\item{.crop_args}{the list of coordinates to use for cropping (default = \code{list(S = 27, N = 72, W = -13, E = 56)})}
}
\value{
a stars object
}
\description{
This function should not be called by the user. It is called internally by
\code{\link[=build_source_stars]{build_source_stars()}}. It reads the content of a NC (NetCDF) file and offer the possibility to
crop and downsample it. Note that NC files we use here as input are already cropped but they
still require to be cropped during reading to make sure the boundaries are defined properly.
}
\examples{
run <- FALSE
if (run) {
  path_NC <- "../ISIMIP_sources/gswp3-w5e5/OBSCLIM/gswp3-w5e5_obsclim_tas_lat27.0to72.0lon-13.0to56.0_daily_2011_2019.nc"
  obsclim_2011_2019 <- load_NC_file(path_NC)
  plot(obsclim_2011_2019[,,,1]) ## print map for day 1
}

}
\seealso{
\code{\link[=load_NC_files]{load_NC_files()}}, \code{\link[=build_source_stars]{build_source_stars()}}
}
