% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNCfiles.R
\name{load_NC_files}
\alias{load_NC_files}
\title{Loads a series of temperature prediction grids stored as NC files}
\usage{
load_NC_files(
  directory_NCfiles,
  downsample = FALSE,
  crop = TRUE,
  .downsampling_args = c(5, 5, 0),
  .crop_args = list(S = 27, N = 72, W = -13, E = 56),
  nb_cores = 2
)
}
\arguments{
\item{directory_NCfiles}{the path to a folder containing subfolders with NC files}

\item{downsample}{a boolean indicating whether to reduce the resolution or not (default = \code{FALSE})}

\item{crop}{a boolean indicating whether to crop or not (default = \code{TRUE})}

\item{.downsampling_args}{arguments for \code{\link[stars:st_downsample]{stars::st_downsample()}} (default = \code{c(5, 5, 0)}, i.e. aggregate space but not time)}

\item{.crop_args}{the list of coordinates to use for cropping (default = \code{list(S = 27, N = 72, W = -13, E = 56)})}

\item{nb_cores}{the number of CPU cores to use (don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}
}
\value{
a stars object
}
\description{
This function should not be called by the user. It is called internally by
\code{\link[=build_source_stars]{build_source_stars()}}. It function calls \code{\link[=load_NC_file]{load_NC_file()}} repeatedly on several NC files and
combine their output. This is required because \href{https://www.isimip.org/}{the Inter-Sectoral Impact Model Intercomparison Project} provides data in slices of years.
}
\details{
This function handles parallel processing but not that this requires a huge amount of memory.
}
\examples{
run <- FALSE
if (run) {
  path_NC_dir <- "../NC/ISIMIP_sources/gswp3-w5e5/OBSCLIM/"
  obsclim <- load_NC_files(path_NC_dir, downsample = TRUE)
  obsclim
 }

}
\seealso{
\code{\link[=load_NC_file]{load_NC_file()}}, \code{\link[=build_source_stars]{build_source_stars()}}
}
