% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{load_physio_datafile}
\alias{load_physio_datafile}
\title{Load a file containing the physiological data}
\usage{
load_physio_datafile(filename)
}
\arguments{
\item{filename}{the name of the data to be imported}
}
\value{
a dataframe
}
\description{
TODO Kseniia: explain logger the structure of the files.
This function should not be directly called by the user.
It is called internally when calling \code{\link[=build_physio_table]{build_physio_table()}}.
}
\details{
Note: for the function to work, the files must be named following the same structure we used.
}
\examples{
filespath <- list.files(system.file("extdata/physio", package = "winteR"), full.names = TRUE)
data_physio <- load_physio_datafile(filespath[1])
head(data_physio)

}
\seealso{
\code{\link[=classify_physio_state]{classify_physio_state()}}, \code{\link[=build_physio_table]{build_physio_table()}}, \code{\link[=clean_physio_table]{clean_physio_table()}}
}
