% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mask_country}
\alias{mask_country}
\title{A mask with country borders for the world}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1627 rows and 1 columns.
}
\usage{
mask_country
}
\description{
A mask with country borders for the world
}
\keyword{datasets}
