% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mask_ocean}
\alias{mask_ocean}
\title{A mask with large water bodies the world}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
mask_ocean
}
\description{
A mask with large water bodies the world
}
\keyword{datasets}
