% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neyuki.R
\name{neyuki}
\alias{neyuki}
\title{Start/End of long-term snow}
\usage{
neyuki(snow, nmin = 10, method = "JMA")
}
\arguments{
\item{snow}{snow depth vector (numeric)}

\item{nmin}{minimum days for long-term snow (int)}

\item{method}{method for calculation (chr; JMA or none)}
}
\value{
logical
}
\description{
Calculate long-term snow span from snow-depth vector
}
