% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oceans_polygons}
\alias{oceans_polygons}
\title{Mask of world oceans}
\format{
A \emph{simple feature collection} object
}
\source{
This object is derived from the package \pkg{rnaturalearth}.
}
\description{
This dataset contains a polygon shapefile that can be used to mask large
bodies of water.
}
\examples{
data("oceans_polygons", package = "winteR")
oceans_polygons
}
\seealso{
\link{lands_polygons} for another mask
}
\keyword{datasets}
