% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{plot_budget_curves}
\alias{plot_budget_curves}
\title{Plot budget curves}
\usage{
plot_budget_curves(
  fit_state,
  fit_MR,
  rangeTa = c(-15, 35),
  Tmirror = 2,
  base_size = 11
)
}
\arguments{
\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{rangeTa}{the range of ambient temperature to consider in plot}

\item{Tmirror}{the ambient temperature around which predictions are mirrored}

\item{base_size}{base font size, given in pts}
}
\value{
a ggplot object
}
\description{
Plot budget curves
}
\examples{
data("fit_torpor", package = "winteR")
data("fit_normo_cauchit", package = "winteR")
plot_budget_curves(fit_state = fit_normo_cauchit, fit_MR = fit_torpor)

}
