% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{plot_budget_panel}
\alias{plot_budget_panel}
\title{Plot the energy budget}
\usage{
plot_budget_panel(
  data_budget,
  y = "g_fat_per_winter",
  threshold_mortality = 27,
  base_size = 11
)
}
\arguments{
\item{data_budget}{a dataframe produced by \code{\link[=compute_budget_df]{compute_budget_df()}}}

\item{y}{a string of characters indicating what y-variable to plot:
"g_fat_per_state", "g_fat_per_day", or "g_fat_per_winter"}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 27)}

\item{base_size}{base font size, given in pts}
}
\value{
a ggplot object
}
\description{
Plot the energy budget
}
\examples{
file_Kharkiv <- paste0(system.file("extdata/weather_real", package = "winteR"),
                       "/Kharkiv_weather_2011_2012.csv")
data_Kharkiv <- build_temp_2years(file_Kharkiv)
Tskin_files <- list.files(system.file("extdata/Tskin", package = "winteR"), full.names = TRUE)
data_Tskin <- build_Tskin_table(Tskin_files)
data_normothermy <- data_Tskin[data_Tskin$Included, ]
fit_normo_cauchit <- spaMM::fitme(Normo ~ Ta + (1|ID), data = data_normothermy,
                                 family = binomial(link = "cauchit"))
data("fit_torpor", package = "winteR")
data_nrg <- compute_budget_df(data_Kharkiv, fit_state = fit_normo_cauchit, fit_MR = fit_torpor)
plot_budget_panel(data_nrg, y = "g_fat_per_state")
plot_budget_panel(data_nrg, y = "g_fat_per_day")
plot_budget_panel(data_nrg, y = "g_fat_per_winter")

}
