% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_fat_map.panel}
\alias{plot_fat_map.panel}
\title{Plot maps showing the variation in winter budget through Europe}
\usage{
plot_fat_map.panel(
  stars1,
  stars2,
  stars3,
  stars4,
  strip_names = NULL,
  threshold_mortality = 27
)
}
\arguments{
\item{stars1}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{stars2}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{stars3}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{stars4}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{strip_names}{the title to be used on top of each map (default, if NULL = \verb{= c("Obs (1901-1930)", "Obs (1989-2018)", "SSP1-2.6 (2070-2099)", "SSP5-8.5 (2070-2099)")})}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 27)}
}
\value{
a ggplot2 object
}
\description{
This function is not used for now and thus not exported.
}
\examples{
run <- FALSE
if (run) {
 readRDS("../NC/stars_winter/gswp3-w5e5_OBSCLIM_winter.rds") |>
  dplyr::select(Budget_winter) |>
  dplyr::filter(year >= 1901, year <= 1930) -> winter_stars_OBS_1901_1930

 readRDS("../NC/stars_winter/gswp3-w5e5_OBSCLIM_winter.rds") |>
   dplyr::select(Budget_winter) |>
   dplyr::filter(year >= 1989, year <= 2018) -> winter_stars_OBS_1989_2018

 reshape_stars_across.models("../NC/stars_winter", SSP = "126") |>
   dplyr::filter(year >= 2070) -> winter_stars_SSP126_2070_2099

 reshape_stars_across.models("../NC/stars_winter", SSP = "585") |>
   dplyr::filter(year >= 2070) -> winter_stars_SSP585_2070_2099

 maps <- plot_fat_map.panel(winter_stars_OBS_1901_1930,
                            winter_stars_OBS_1989_2018,
                            winter_stars_SSP126_2070_2099,
                            winter_stars_SSP585_2070_2099)
 showtext::showtext_opts(dpi = 300)
}

}
