% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{plot_flightpower}
\alias{plot_flightpower}
\title{Plot power available for flight and power required for flight as a function of body mass}
\usage{
plot_flightpower(base_size = 11)
}
\arguments{
\item{base_size}{base font size, given in pts}
}
\value{
a ggplot
}
\description{
Plot power available for flight and power required for flight as a function of body mass
}
\examples{
\dontrun{
plot_flightpower()
}

}
