% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kharkiv.R
\name{plot_Kharkiv_temp}
\alias{plot_Kharkiv_temp}
\title{Plot Kharkiv temperature data}
\usage{
plot_Kharkiv_temp(data, base_size = 11, colour_var = NULL, temp_threshold = 7)
}
\arguments{
\item{data}{the dataframe created by \code{\link[=build_Kharkiv_table]{build_Kharkiv_table()}}}

\item{base_size}{base font size, given in pts}

\item{colour_var}{(optional) the unquotted name of a variable in \code{data} that defines the colour in the plot}

\item{temp_threshold}{the approximate temperature below which insects do not fly}
}
\value{
a ggplot object
}
\description{
Plot Kharkiv temperature data
}
\examples{
filepath <- list.files(system.file("extdata/weather_real", package = "winteR"),
                       full.names = TRUE)[1]
data_Kharkiv <- build_Kharkiv_table(filepath)
plot_Kharkiv_temp(data_Kharkiv)


}
