% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{plot_physio_table}
\alias{plot_physio_table}
\title{Plot the physiological data}
\usage{
plot_physio_table(data)
}
\arguments{
\item{data}{the dataframe produced by \code{\link[=build_physio_table]{build_physio_table()}}}
}
\value{
a plot
}
\description{
Plot the physiological data
}
\examples{
files_to_do <- list.files(system.file("extdata/physio", package = "winteR"), full.names = TRUE)
data_physio <- build_physio_table(files_to_do)
plot_physio_table(data_physio)

}
