% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_suitability_map}
\alias{plot_suitability_map}
\title{Plot maps showing the variation in hibernation suitability budget through Europe}
\usage{
plot_suitability_map(
  stars_tbl,
  scenario = "SSP126",
  starsname = "stars_avg",
  varname = "freq_suitability_pct",
  polygons = NULL,
  base_size = 9,
  legend_position = "left"
)
}
\arguments{
\item{stars_tbl}{a tibble of stars objects produced by \code{\link[=build_suitability_stars]{build_suitability_stars()}}}

\item{scenario}{the scenario to be selected (\code{"SSP126"}, \code{"SSP245"}, \code{"SSP370"} or \code{"SP585"})}

\item{starsname}{the name of the column containing the stars objects to be used (in the tibble produced by \code{\link[=build_suitability_stars]{build_suitability_stars()}})}

\item{varname}{the quoted name of the variable to retain from each stars object (default depends on the function)}

\item{polygons}{a list of sf (MULTI)POLYGON object representing the distribution of the species}

\item{base_size}{base font size, given in pts}

\item{legend_position}{the value for the ggplot argument "legend.position" used in \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
a ggplot2 object
}
\description{
This function produces the figure 6.
}
\examples{
# see ?winterR

}
