% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoreg.R
\name{plot_TaTskin_data}
\alias{plot_TaTskin_data}
\title{Plot ambient vs skin temperature}
\usage{
plot_TaTskin_data(
  fit_MR,
  data_MR,
  rangeTa = c(-5, 35),
  rangeTskin = c(0, 40),
  base_size = 11
)
}
\arguments{
\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{data_MR}{the data used to fit the thermoregulatory curves as produced by \code{\link[=build_MR_table]{build_MR_table()}}}

\item{rangeTa}{the range of ambient temperature to consider in plot}

\item{rangeTskin}{the range of skin temperature to consider in plot}

\item{base_size}{base font size, given in pts}
}
\value{
a ggplot object
}
\description{
Plot ambient vs skin temperature
}
\examples{
filepath <- paste0(system.file("extdata/thermoreg", package = "winteR"),
                   "/metabolic_rate.txt")
data_MR <- build_MR_table(filepath)
plot_TaTskin_data(fit_MR = fit_torpor, data_MR = data_MR)

}
