% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_trends.R
\name{plot_time_trends}
\alias{plot_time_trends}
\title{Plot how a variable connected to the hibernation niche changes over time}
\usage{
plot_time_trends(
  winters_stats_df,
  window_length_smoothing = 10,
  varname = "Latitude_median",
  vartype = c("latitude", "area"),
  y_title = "\\n Median latitude of suitable hibernation areas",
  scenarios = c("OBSCLIM", "SSP126", "SSP585"),
  .fill_values = c("NA", "blue", "orange"),
  .color_values = c("black", "blue", "orange"),
  .legend_position = "none",
  base_size = 9
)
}
\arguments{
\item{winters_stats_df}{a dataframe of winter statistics produced by \code{\link[=summarise_info_winter.stars.all]{summarise_info_winter.stars.all()}}}

\item{window_length_smoothing}{the width of the window used for smoothing via \code{\link[caTools:runmean]{caTools::runmean()}}}

\item{varname}{the quoted name of the variable to retain from each stars object (default depends on the function)}

\item{vartype}{the type of information for the y variable: \code{"latitude"} or \code{"area"}}

\item{y_title}{the title for the y axis}

\item{scenarios}{a vector of the scenarios to retain (default = \code{c("OBSCLIM", "SSP126", "SSP585")})}

\item{.fill_values}{a vector of colors for the \code{fill} variable (default = \code{c("NA", "blue", "orange")})}

\item{.color_values}{a vector of colors for the \code{color} variable (default = \code{c("NA", "blue", "orange")})}

\item{.legend_position}{the value for the ggplot argument "legend.position" used in \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{base_size}{base font size, given in pts}
}
\value{
a ggplot2 object
}
\description{
Plot how a variable connected to the hibernation niche changes over time
}
\examples{
## see ?winteR

}
