% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tskin_normothermy.R
\name{plot_Tskin_fit}
\alias{plot_Tskin_fit}
\title{Title}
\usage{
plot_Tskin_fit(fit, rangeTa = c(-5, 35), Tmirror = 2, base_size = 11)
}
\arguments{
\item{fit}{a model fitted with \code{\link[spaMM:fitme]{spaMM::fitme()}}}

\item{rangeTa}{the range of ambient temperature to consider in predictions}

\item{Tmirror}{the ambient temperature around which predictions are mirrored}

\item{base_size}{base font size, given in pts}
}
\value{
a plot
}
\description{
Title
}
\examples{
#TODO

}
