% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_series.R
\name{plot_winter_temp2years}
\alias{plot_winter_temp2years}
\title{Plot temperature and winter data}
\usage{
plot_winter_temp2years(
  data_temp,
  base_size = 11,
  temp_threshold = 7,
  split_summer = "07-01",
  min_days_trigger_winter = 14,
  roost_insulation_dTa = 5
)
}
\arguments{
\item{data_temp}{a dataframe with at least columns \code{Date} and \code{Temp}, such as those produced by \code{\link[=build_temp_2years]{build_temp_2years()}}}

\item{base_size}{base font size, given in pts}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}
}
\value{
a ggplot object
}
\description{
Plot temperature and winter data
}
\examples{
file_Kharkiv <- paste0(system.file("extdata/weather_real", package = "winteR"),
                       "/Kharkiv_weather_2011_2012.csv")
data_Kharkiv <- build_temp_2years(file_Kharkiv)
plot_winter_temp2years(data_Kharkiv)


}
