% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_series.R
\name{recode_freq_pct}
\alias{recode_freq_pct}
\title{Helper function to turn proportion into categories of percentages}
\usage{
recode_freq_pct(vec_Proportion)
}
\arguments{
\item{vec_Proportion}{a vector of proportion}
}
\value{
a vector
}
\description{
Helper function to turn proportion into categories of percentages
}
\examples{
recode_freq_pct(seq(0, 1, 0.01))

}
