% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_series.R
\name{recode_year_decade}
\alias{recode_year_decade}
\title{Helper function to turn years into decades}
\usage{
recode_year_decade(vec_Year)
}
\arguments{
\item{vec_Year}{a vector of years}
}
\value{
a vector
}
\description{
Helper function to turn years into decades
}
\examples{
recode_year_decade(c(1910, 2022, 2099))

}
