% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{reshape_stars_across.models}
\alias{reshape_stars_across.models}
\title{Loads, combine and merge winter stars across forcing models}
\usage{
reshape_stars_across.models(
  directory_stars,
  SSP,
  varname = "Budget_winter",
  flatten = TRUE
)
}
\arguments{
\item{directory_stars}{the destination location for the rds files storing the stars object}

\item{SSP}{the Shared Socioeconomic Pathways to consider ("126", "245", or "585")}

\item{varname}{the quotted name of the variable to retain from each winter stars (default = \code{"Budget_winter"})}

\item{flatten}{whether to average the winter stars across forcing models (default = \code{TRUE})}
}
\value{
a stars object
}
\description{
Loads, combine and merge winter stars across forcing models
}
\examples{
run <- FALSE
if (run) {
 reshape_stars_across.models(directory_stars = "../NC/stars_winter", SSP = "126")
 }

}
