% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{summarise_info_winter.stars}
\alias{summarise_info_winter.stars}
\title{Summarise winter information}
\usage{
summarise_info_winter.stars(stars_object, mask = NULL)
}
\arguments{
\item{stars_object}{a stars object}

\item{mask}{a MULTIPOLYGON object to be used as a mask}
}
\value{
a dataframe
}
\description{
This function should not be called by the user. It is internally called by \code{\link[=summarise_info_winter.stars.all]{summarise_info_winter.stars.all()}}.
}
\examples{
run <- FALSE
if (run) {
 test_stars <- readRDS("../NC/stars_winter/gfdl-esm4_SSP126_winter.rds")
 #test_stars_small <- dplyr::filter(test_stars, year > 2080, x < -10, y > 70)
 data(lands_polygons, package = "winteR")
 test <- summarise_info_winter.stars(test_stars, mask = lands_polygons)
 }

}
