% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winteR-package.R
\docType{package}
\name{winteR-package}
\alias{winteR-package}
\alias{winteR}
\title{ADD TITLE PAPER HERE}
\description{
The goal of winteR is to reproduce the results from the paper entitled "XXX" by Kravchenko et al.
}
\details{
For reproducing the results of our paper, just follow the code presented in the section "Examples" below.
}
\examples{
1+1
}
\references{
ADD REF PAPER HERE
}
\keyword{package}
