% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_overs.R
\name{winter_doy}
\alias{winter_doy}
\title{Get continuous day-of-year index}
\usage{
winter_doy(x, end = 3)
}
\arguments{
\item{x}{a date-time object}

\item{end}{month of winter end}
}
\value{
doy
}
\description{
DOY index continues to the following Jan-Feb-Mar.
}
\examples{
x <- lubridate::ymd(c("1999-10-30", "1999-01-03", "2000-10-30", "2001-01-03", "2000-09-01", "20o0-01-0l"))
winter_doy(x)
}
