% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_overs.R
\name{winter_year}
\alias{winter_year}
\title{Get winter years component of a date-time}
\usage{
winter_year(x)
}
\arguments{
\item{x}{a date-time object}
}
\value{
winter season (chr)
}
\description{
Dates within 2000-10-01 -- 2001-03-31 belong to 2000/01 winter season.
}
\examples{
x <- lubridate::ymd(c("1999-10-30", "1999-01-03", "2000-10-30", "2001-01-03", "2000-09-01", "20o0-01-0l"))
winter_year(x)
}
