% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{arguments}
\alias{arguments}
\title{Definition of the function arguments}
\arguments{
\item{base_size}{base font size, given in pts}

\item{clean}{a boolean indicating if the data need to be filter by \code{\link[=clean_Tskin_table]{clean_Tskin_table()}} or not (default = \code{TRUE})}

\item{.cluster}{the registered default cluster (set internally)}

\item{crop}{a boolean indicating whether to crop or not (default = \code{TRUE})}

\item{.crop_args}{the list of coordinates to use for cropping (default = \code{list(S = 27, N = 72, W = -13, E = 56)})}

\item{data_budget}{a dataframe produced by \code{\link[=compute_budget_df]{compute_budget_df()}}}

\item{data_MR}{the data used to fit the thermoregulatory curves as produced by \code{\link[=build_MR_table]{build_MR_table()}}}

\item{data_temp}{a dataframe with at least columns \code{Date} and \code{Temp}, such as those produced by \code{\link[=build_temp_2years]{build_temp_2years()}}}

\item{data_Tskin}{a file produced by \code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}}

\item{directory_NCfiles}{the path to a folder containing subfolders with NC files}

\item{directory_stars}{the destination location for the rda files storing the stars object}

\item{downsample}{a boolean indicating whether to reduce the resolution or not (default = \code{FALSE})}

\item{.downsampling_args}{arguments for \code{\link[stars:st_downsample]{stars::st_downsample()}} (default = \code{c(5, 5, 0)}, i.e. aggregate space but not time)}

\item{filename}{the path to the file from which the data are to be imported}

\item{filenames}{a vector of paths to the files from which the data are to be imported}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{nb_cores}{the number of CPU cores to use (don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}

\item{nb_cores_location}{the number of CPU cores to use to parallelize the same task across multiple locations (don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}

\item{nb_cores_year}{the number of CPU cores to use to parallelize the same task across multiple years (don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}

\item{metadirectory_NCfiles}{a folder containing one subfolder per climate model, themselves containing one subfolder per SSP scenarios, themsevles containing the NC files}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}

\item{rangeTa}{the range of ambient temperature to consider in plot}

\item{rangeTskin}{the range of skin temperature to consider in plot}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{stars_object}{a stars object containing daily temperatures}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 24)}

\item{Tmirror}{the ambient temperature around which predictions are mirrored}

\item{vec_Temp}{a vector of ambient temperatures over 2 consecutive years}

\item{vec_Dates}{a vector of dates over 2 consecutive years}

\item{y}{a string of characters indicating what y-variable to plot:
"g_fat_per_state", "g_fat_per_day", or "g_fat_per_winter"}

\item{year_start}{a numeric scalar indicating the first year of the winter to consider}
}
\description{
Here is the list of most function arguments used in the package.
The arguments not contained in this list are those for which the exact meaning depends on the context.
}
