% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{arguments}
\alias{arguments}
\title{Definition of the function arguments}
\arguments{
\item{base_size}{base font size, given in pts}

\item{clean}{a boolean indicating if the data need to be filter by \code{\link[=clean_Tskin_table]{clean_Tskin_table()}} or not (default = \code{TRUE})}

\item{.cluster}{the registered default cluster (set internally)}

\item{.color_values}{a vector of colors for the \code{color} variable (default = \code{c("NA", "blue", "orange")})}

\item{crop}{a boolean indicating whether to crop or not (default = \code{TRUE})}

\item{.crop_args}{the list of coordinates to use for cropping (default = \code{list(S = 27, N = 72, W = -13, E = 56)})}

\item{data_budget}{a dataframe produced by \code{\link[=compute_budget_df]{compute_budget_df()}}}

\item{data_MR}{the data used to fit the thermoregulatory curves as produced by \code{\link[=build_MR_table]{build_MR_table()}}}

\item{data_suitability}{a dataframe with at least columns \code{Year} and \code{Suitable_winter}}

\item{data_temp}{a dataframe with at least columns \code{Date} and \code{Temp}, such as those produced by \code{\link[=build_temp_2years]{build_temp_2years()}}}

\item{data_Tskin}{a file produced by \code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}}

\item{directory_NCfiles}{the path to a folder containing subfolders with NC files}

\item{directory_stars}{the destination location for the rds files storing the stars object}

\item{downsample}{a boolean indicating whether to reduce the resolution or not (default = \code{FALSE})}

\item{.downsampling_args}{arguments for \code{\link[stars:st_downsample]{stars::st_downsample()}} (default = \code{c(5, 5, 0)}, i.e. aggregate space but not dates)}

\item{filename}{the path to the file from which the data are to be imported}

\item{filenames}{a vector of paths to the files from which the data are to be imported}

\item{.fill_values}{a vector of colors for the \code{fill} variable (default = \code{c("NA", "blue", "orange")})}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{flatten}{whether to average the winter stars across forcing models (default = \code{TRUE})}

\item{fn}{a function to be applied}

\item{huddling_factor}{a factor multiplied to the metabolic rate, to account for huddling, when individuals are thermoregulating (default = 0.5)}

\item{IUCN_polygon}{a sf (MULTI)POLYGON object representing the IUCN distribution of the specoes}

\item{lapply_pkg}{the R package used to implement a \code{lapply()} kind of function (default = "pbmcapply"; other possibilities are "parallel" and "base")}

\item{lat}{a latitude}

\item{lat_ref}{a latitude of reference}

\item{.legend_position}{the value for the ggplot argument "legend.position" used in \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{name_bool_var}{the quoted name of a boolean variable to use to filter cells in the grid (default is "Survive")}

\item{nb_cores}{the number of CPU cores to use (for Linux or MacOS only, not Windows, and don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}

\item{mask}{a MULTIPOLYGON object to be used as a mask}

\item{max_Year}{the maximum year (in numeric format) to consider in the factor}

\item{metadirectory_NCfiles}{a folder containing one subfolder per climate model, themselves containing one subfolder per SSP scenarios, themselves containing the NC files}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}

\item{min_Year}{the minimum year (in numeric format) to consider in the factor}

\item{min_years_trigger_suitability}{the minimum number of suitable winters to trigger the establishment or the disappearance of bats in a region}

\item{rangeTa}{the range of ambient temperature to consider in plot}

\item{rangeTskin}{the range of skin temperature to consider in plot}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}

\item{scenario}{the scenario to be selected (\code{"SSP126"}, \code{"SSP245"}, \code{"SSP370"} or \code{"SP585"})}

\item{scenarios}{a vector of the scenarios to retain (default = \code{c("OBSCLIM", "SSP126", "SSP585")})}

\item{signed}{whether the output value should be signed or absolute (default = \code{TRUE}, which implies it is signed)}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{SSP}{the Shared Socioeconomic Pathways to consider ("126", "245", or "585")}

\item{stars1}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{stars2}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{stars3}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{stars4}{a stars object produced by \code{\link[=compute_budget_stars]{compute_budget_stars()}}}

\item{stars_object}{a stars object}

\item{stars_tbl}{a tibble of stars objects produced by \code{\link[=build_suitability_stars]{build_suitability_stars()}}}

\item{starsname}{the name of the column containing the stars objects to be used (in the tibble produced by \code{\link[=build_suitability_stars]{build_suitability_stars()}})}

\item{strip_names}{the title to be used on top of each map (default, if NULL = \verb{= c("Obs (1901-1930)", "Obs (1989-2018)", "SSP1-2.6 (2070-2099)", "SSP5-8.5 (2070-2099)")})}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 27)}

\item{Tmirror}{the ambient temperature around which predictions are mirrored}

\item{varname}{the quoted name of the variable to retain from each stars object (default depends on the function)}

\item{vartype}{the type of information for the y variable: \code{"latitude"} or \code{"area"}}

\item{vec_Dates}{a vector of dates over 2 consecutive years}

\item{vec_Proportion}{a vector of proportion}

\item{vec_Suitable_winter}{a vector of boolean indicating if a hibernation season was suitable (\code{TRUE}) or not (\code{FALSE}) (nb: \code{FALSE} includes lack of hibernation season as well as hibernation season where the budget is not sufficient to survive)}

\item{vec_Temp}{a vector of ambient temperatures over 2 consecutive years}

\item{vec_Year}{a vector of years}

\item{window_length_smoothing}{the width of the window used for smoothing via \code{\link[caTools:runmean]{caTools::runmean()}}}

\item{winters_stats_df}{a dataframe of winter statistics produced by \code{\link[=summarise_info_winter.stars.all]{summarise_info_winter.stars.all()}}}

\item{y}{a string of characters indicating what y-variable to plot:
"g_fat_per_state", "g_fat_per_day", or "g_fat_per_winter"}

\item{y_title}{the title for the y axis}

\item{year}{a numeric scalar indicating the year of the winter to consider}

\item{year_start}{a numeric scalar indicating the first year of the winter to consider}
}
\description{
Here is the list of most function arguments used in the package.
The arguments not contained in this list are those for which the exact meaning depends on the context.
}
