% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{compute_budget_stars}
\alias{compute_budget_stars}
\title{Compute summary statistics for all winters from a stars object}
\usage{
compute_budget_stars(
  stars_object,
  fit_state,
  fit_MR,
  roost_insulation_dTa = 5,
  huddling_factor = 0.5,
  temp_threshold = 7,
  split_summer = "07-01",
  min_days_trigger_winter = 14,
  threshold_mortality = 27,
  nb_cores = 1,
  lapply_pkg = "pbmcapply"
)
}
\arguments{
\item{stars_object}{a stars object}

\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}

\item{huddling_factor}{a factor multiplied to the metabolic rate, to account for huddling, when individuals are thermoregulating (default = 0.5)}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 27)}

\item{nb_cores}{the number of CPU cores to use (for Linux or MacOS only, not Windows, and don't use too many cores otherwise, you may reach the RAM limit and it won't work. If error, do reduce)}

\item{lapply_pkg}{the R package used to implement a \code{lapply()} kind of function (default = "pbmcapply"; other possibilities are "parallel" and "base")}
}
\value{
a stars object with energy budget information
}
\description{
This function calls \code{\link[=compute_budget_summarystats]{compute_budget_summarystats()}} on all winters and across all locations
contained in a stars object. It is an extremely computationally intensive step. Note that the
progress messages are more informative outside RStudio.
}
\examples{
run <- FALSE
if (run) {
 data("fit_torpor", package = "winteR")
 data("fit_normo_cauchit", package = "winteR")
 test_stars <- readRDS("../NC/stars/gfdl-esm4_SSP126.rds")
 test_stars_small <- dplyr::filter(test_stars, time < as.Date("2017-01-01"), x < -10, y > 70)
 test <- compute_budget_stars(test_stars_small,
                              fit_state = fit_normo_cauchit, fit_MR = fit_torpor,
                              nb_cores = 30)
 test
 }

}
