% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{compute_budget_stars1year}
\alias{compute_budget_stars1year}
\title{Compute budget for one winter from a stars object}
\usage{
compute_budget_stars1year(
  year_start,
  stars_object,
  fit_state,
  fit_MR,
  roost_insulation_dTa = 5,
  huddling_factor = 0.5,
  temp_threshold = 7,
  split_summer = "07-01",
  min_days_trigger_winter = 14,
  threshold_mortality = 27
)
}
\arguments{
\item{year_start}{a numeric scalar indicating the first year of the winter to consider}

\item{stars_object}{a stars object}

\item{fit_state}{a fitted model predicting the probability to be in normothermy}

\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}

\item{roost_insulation_dTa}{the increase in temperature in the roost, compared to outside (default = 5)}

\item{huddling_factor}{a factor multiplied to the metabolic rate, to account for huddling, when individuals are thermoregulating (default = 0.5)}

\item{temp_threshold}{the approximate temperature below which insects do not fly}

\item{split_summer}{the day of mid-summer (default = \code{'07-01'})}

\item{min_days_trigger_winter}{the minimum number of days for which the temperature should be below \code{temp_threshold} to enter winter}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 27)}
}
\value{
a stars object with energy budget information
}
\description{
This function should not be called by the user. It is called internally by
\code{\link[=compute_budget_stars]{compute_budget_stars()}}.
}
\examples{
run <- FALSE
if (run) {
  test_stars <- readRDS("../NC/stars/gfdl-esm4_SSP126.rds")
  test <- compute_budget_stars1year(year_start = 2020, test_stars,
                                    fit_state = fit_normo_cauchit, fit_MR = fit_torpor)
 }

}
