% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_fat_map}
\alias{plot_fat_map}
\title{Plot maps showing the variation in winter budget through Europe}
\usage{
plot_fat_map(
  stars_object,
  polygons = NULL,
  threshold_mortality = 27,
  base_size = 11
)
}
\arguments{
\item{stars_object}{a stars object}

\item{polygons}{a list of sf (MULTI)POLYGON object representing the distribution of the species}

\item{threshold_mortality}{the maximal amount of fat consumed before mortality occurs (default = 27)}

\item{base_size}{base font size, given in pts}
}
\value{
a ggplot2 object
}
\description{
This function produces the figure 4.
}
\examples{
# see ?winterR

}
