% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoreg.R
\name{plot_MR_fit}
\alias{plot_MR_fit}
\title{Plot thermoregulatory curves}
\usage{
plot_MR_fit(fit)
}
\arguments{
\item{fit}{a model fit produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}
}
\value{
a ggplot object
}
\description{
Plot thermoregulatory curves
}
\examples{
plot_MR_fit(fit_torpor)

}
