% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoreg.R
\name{plot_TaTskin_data}
\alias{plot_TaTskin_data}
\title{Plot ambient vs skin temperature}
\usage{
plot_TaTskin_data(
  data,
  rangeTa = c(-5, 35),
  rangeTskin = c(0, 40),
  base_size = 11
)
}
\arguments{
\item{data}{the data used to fit the thermoregulatory curves as produced by \code{\link[=build_MR_datafile]{build_MR_datafile()}}}

\item{rangeTa}{the range of ambient temperature to consider in plot}

\item{rangeTskin}{the range of skin temperature to consider in plot}

\item{base_size}{base font size, given in pts}
}
\value{
a ggplot object
}
\description{
Plot ambient vs skin temperature
}
\examples{
Tskin_files <- list.files(system.file("extdata/Tskin", package = "winteR"), full.names = TRUE)
data_Tskin <- build_Tskin_table(Tskin_files)
plot_TaTskin_data(data_Tskin)

}
