% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tskin_normothermy.R
\name{plot_Tskin_fit}
\alias{plot_Tskin_fit}
\title{Plot a fitted model between ambient temperature and normothermy}
\usage{
plot_Tskin_fit(fit, rangeTa = c(-5, 35), Tmirror = 2, base_size = 11)
}
\arguments{
\item{fit}{a model fitted with \code{\link[spaMM:fitme]{spaMM::fitme()}}}

\item{rangeTa}{the range of ambient temperature to consider in plot}

\item{Tmirror}{the ambient temperature around which predictions are mirrored}

\item{base_size}{base font size, given in pts}
}
\value{
a plot
}
\description{
This function plots the predictions from a model fitting the relationship between ambient
temperature and normothermy. It assumes that the predictions are mirrored around a particular
ambient temperature value (default = 2 degrees C). We use this function to create Fig 1 and Fig
S1 in the paper.
}
\examples{
#See ?winteR

}
