% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tskin_normothermy.R
\name{plot_Tskin_table}
\alias{plot_Tskin_table}
\title{Plot the physiological data on skin temperature}
\usage{
plot_Tskin_table(data)
}
\arguments{
\item{data}{the dataframe produced by \code{\link[=build_Tskin_table]{build_Tskin_table()}}}
}
\value{
a plot
}
\description{
Plot the physiological data on skin temperature
}
\examples{
files_to_do <- list.files(system.file("extdata/Tskin", package = "winteR"), full.names = TRUE)
data_physio <- build_Tskin_table(files_to_do)
plot_Tskin_table(data_physio)

}
