% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tskin_normothermy.R
\name{plot_Tskin_table}
\alias{plot_Tskin_table}
\title{Plot the physiological data on skin temperature}
\usage{
plot_Tskin_table(data_Tskin, base_size = 11)
}
\arguments{
\item{data_Tskin}{a file produced by \code{\link[=load_Tskin_datafile]{load_Tskin_datafile()}}}

\item{base_size}{base font size, given in pts}
}
\value{
a plot
}
\description{
This function plots the data created with \code{\link[=build_Tskin_table]{build_Tskin_table()}}.
}
\examples{
files_to_do <- list.files(system.file("extdata/Tskin", package = "winteR"), full.names = TRUE)
data_Tskin <- build_Tskin_table(files_to_do)
plot_Tskin_table(data_Tskin)

}
