% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleStars.R
\name{summarise_info_winter.stars}
\alias{summarise_info_winter.stars}
\title{Summarise winter info}
\usage{
summarise_info_winter.stars(stars_object, mask = NULL)
}
\arguments{
\item{stars_object}{a stars object containing daily temperatures}

\item{mask}{a MULTIPOLYGON object to be used as a mask}
}
\value{
a stars object with a single layer
}
\description{
This function summarises the information contained in a stars object produced with \code{\link[=compute_budget_stars]{compute_budget_stars()}}.
}
\examples{
run <- FALSE
if (run) {
 test_stars <- readRDS("../NC/stars_winter/gfdl-esm4_SSP126_winter.rds")
 #test_stars_small <- dplyr::filter(test_stars, year > 2080, x < -10, y > 70)
 test <- summarise_info_winter.stars(test_stars, mask = mask_country)
 }

}
