% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoreg.R
\name{summarise_MR_fit}
\alias{summarise_MR_fit}
\title{Extract parameters from model fitted with torpor}
\usage{
summarise_MR_fit(fit_MR)
}
\arguments{
\item{fit_MR}{a fitted model predicting the metabolic rate in KJh^-1 produced by \code{\link[torpor:tor_fit]{torpor::tor_fit()}}}
}
\value{
a dataframe
}
\description{
Extract parameters from model fitted with torpor
}
\examples{
summarise_MR_fit(fit_torpor)

}
