% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter-package.R
\docType{package}
\name{winter-package}
\alias{winter}
\alias{winter-package}
\title{winter: Cover Time Series with Carry Over}
\description{
A bundle of functions to deal with time series data with carry over to the following year.
}
\author{
\strong{Maintainer}: Keach Murakami \email{keach.murakami@gmail.com} (0000-0001-8150-9535)

}
\keyword{internal}
